dom/tests/mochitest/general/test_bug1077002.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1077002
-->
<head>
  <title>Test for Bug 1077002</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank"
    href="https://bugzilla.mozilla.org/show_bug.cgi?id=1077002">Mozilla Bug 1077002</a>
<p id="display"></p>
<div id="content">
  <iframe id="frame" style="height:100px; width:100px; border:0"></iframe>
  <div id="status" style="display: none"></div>
</div>
<pre id="test">
<script type="application/javascript;version=1.7">

/** Test for Bug 1077002 **/

try {
    // NB: This test runs in e10s only. In e10s showModalDialog should only
    // ever throw NS_ERROR_NOT_AVAILABLE.
    window.showModalDialog("http://example.org");
} catch (e) {
    is(e.name, "NS_ERROR_NOT_AVAILABLE", "throw the correct error message");
}

</script>
</pre>
</body>
</html>
